services/users_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
pydantic[email]==2.5.3


